ai/requirements.txt
flask==3.0.0
prophet==1.1.5
pandas==2.1.4
redis==5.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
numpy==1.26.2